+++ src/main/java/ru/mail/polis/homework/exception/ConnectionException.java
@@ -1,0 +1,7 @@
+package ru.mail.polis.homework.exception;
+
+class ConnectionException extends Exception {
+    public ConnectionException(String message) {
+        super(message);
+    }
+}
