+++ src/main/java/ru/mail/polis/homework/exception/NoEnergyException.java
@@ -1,0 +1,7 @@
+package ru.mail.polis.homework.exception;
+
+class NoEnergyException extends Exception {
+    public NoEnergyException(String message) {
+        super(message);
+    }
+}
