+++ src/main/java/ru/mail/polis/homework/exception/RobotManager.java
@@ -5,7 +5,7 @@ public class RobotManager implements RobotConnectionManager {
     public RobotConnection getConnection(int robotId) throws ConnectionException, NoEnergyException {
         Robot robot = new Robot(10, 0.5, robotId);
         Connection connection = new Connection(robot, robot.connectionStatus());
-        if (connection.getConnectionStatus() && connection.getID() == robotId) {
+        if (connection.getID() == robotId) {
             return connection;
         } else {
             throw new ConnectionException("Connection error!");
